Refactor Gift Card List Top Alert
